big update 1
vẫn chưa xử lý pgp trong addon group...

+++ app/Http/Controllers/HomeController.php
@@ -20,7 +20,8 @@ use App\GroupUser;
 use App\Share;
 
 use Hash;
-use App\Mail\SendMailable;
+
+use App\Mail\PotentialUser;
 
 
 class HomeController extends Controller
@@ -75,6 +76,12 @@ class HomeController extends Controller
         return view('page.accounts', compact('accounts'));
     }
 
+    public function getAccount(Request $request)
+    {
+        $account = Account::where("id", $request->id)->first();
+        return $account;
+    }
+
     public function addAccount(Request $request)
     {
         $account = new Account;
@@ -87,8 +94,8 @@ class HomeController extends Controller
         // Nối id tới secret ứng mỗi user khác nhau
         $secret = new Secret;
         $user = Auth::user();
-        $secret->user_id = $user->id;
-        $secret->account_id = $account->id;
+        $secret->owner_id = $user->id;
+        $secret->asset_id = $account->id;
         // TODO: encrypt OpenGPG
         $secret->data = $request->cipher;
         $secret->save();
@@ -112,7 +119,7 @@ class HomeController extends Controller
         $acc->save();
 
         if ($request->cipher) {
-            $secret = Secret::where('account_id', $account_id)->first();
+            $secret = Secret::where('asset_id', $account_id)->first();
             $secret->data = $request->cipher;
             $secret->save();
         }
@@ -132,7 +139,7 @@ class HomeController extends Controller
         $acc = Account::find($account_id);
         $acc->delete();
 
-        $secret = Secret::where('account_id', $account_id)->first();
+        $secret = Secret::where('asset_id', $account_id)->first();
         $secret->delete();
 
         $accounts = Auth::user()->account()->get();
@@ -145,29 +152,29 @@ class HomeController extends Controller
     }
     
     public function shareAccount(Request $request){
-        $user = User::where('email',$request->email)->first();
+        $user = User::where('email', $request->email)->first();
         $account_id = $request->id;
         $account = Account::find($account_id);
-        $secret = Secret::where('account_id',$account->id)->first();
+        $secret = Secret::where('asset_id',$account->id)->first();
         // co thi tra ve publickey shared user
         if($user) {
             $newAccount = $account->replicate();
             $newAccount->save();
 
             $newSecret = $secret->replicate();
-            $newSecret->user_id = $user->id;
-            $newSecret->account_id = $newAccount->id;
+            $newSecret->owner_id = $user->id;
+            $newSecret->asset_id = $newAccount->id;
             $newSecret->data = "";
             $newSecret->save();
 
             $share = new Share;
             $share->asset_id = $newAccount->id;
             $share->user_id = $user->id;
-            $share->owner_id = Auth::user()->id;
+            $share->shared_by = Auth::user()->id;
             $share->comment = $request->comment;
             $share->save();
 
-            $sharedkey = PGPkey::where('user_id',$user->id)
+            $sharedkey = PGPkey::where('owner_id', $user->id)
                                 ->where('type','6')
                                 ->first();
             
@@ -181,38 +188,41 @@ class HomeController extends Controller
         }
         else {
             // khong co thi gui mail marketing
+            Mail::to($request->email)->send(new PotentialUser(Auth::user(), $account, "account"));
+
             return response()->json([
                 'success' => true,
                 // TODO: lang this message
-                'message' => 'Người dùng chưa đăng ký'
-            ]);
+                'message' => 'Người nhận chưa đăng ký dịch vụ',
+                'detail' => 'Email thông báo đã được gửi tới người nhận, bạn vẫn cần gửi lại thông tin sau khi người dùng đăng ký dịch vụ thành công.'
+            ],500);
         }
     }
         
     public function shareNote(Request $request){
-        $user = User::where('email',$request->email)->first();
+        $user = User::where('email', $request->email)->first();
         $note_id = $request->id;
         $note = Note::find($note_id);
-        $secret = Secret::where('note_id',$note->id)->first();
+        $secret = Secret::where('asset_id', $note->id)->first();
         // co thi tra ve publickey shared user
         if($user) {
             $newnote = $note->replicate();
             $newnote->save();
 
             $newSecret = $secret->replicate();
-            $newSecret->user_id = $user->id;
-            $newSecret->note_id = $newnote->id;
+            $newSecret->owner_id = $user->id;
+            $newSecret->asset_id = $newnote->id;
             $newSecret->data = "";
             $newSecret->save();
 
             $share = new Share;
             $share->asset_id = $newnote->id;
             $share->user_id = $user->id;
-            $share->owner_id = Auth::user()->id;
+            $share->shared_by = Auth::user()->id;
             $share->comment = $request->comment;
             $share->save();
 
-            $sharedkey = PGPkey::where('user_id',$user->id)
+            $sharedkey = PGPkey::where('owner_id',$user->id)
                                 ->where('type','6')
                                 ->first();
             
@@ -226,11 +236,14 @@ class HomeController extends Controller
         }
         else {
             // khong co thi gui mail marketing
+            Mail::to($request->email)->send(new PotentialUser(Auth::user(), $note, "note"));
+
             return response()->json([
                 'success' => true,
                 // TODO: lang this message
-                'message' => 'Người dùng không tồn tại'
-            ]);
+                'message' => 'Người nhận chưa đăng ký dịch vụ',
+                'detail' => 'Email thông báo đã được gửi tới người nhận, bạn vẫn cần gửi lại thông tin sau khi người dùng đăng ký dịch vụ thành công.'
+            ],500);
         }
     }
 
@@ -248,7 +261,7 @@ class HomeController extends Controller
     public function getPassword(Request $request)
     {
         $account_id = $request->id;
-        $secret = Secret::where('account_id', $account_id)->first();
+        $secret = Secret::where('asset_id', $account_id)->first();
 
         return response()->json([
             'success' => true,
@@ -264,10 +277,16 @@ class HomeController extends Controller
         return view('page.notes',compact('notes'));
     }
 
+    public function getNote(Request $request)
+    {
+        $note = Note::where("id", $request->id)->first();
+        return $note;
+    }
+
     public function getNoteContent(Request $request)
     {
         $note_id = $request->id;
-        $secret = Secret::where('note_id', $note_id)->first();
+        $secret = Secret::where('asset_id', $note_id)->first();
 
         return response()->json([
             'success' => true,
@@ -286,8 +305,8 @@ class HomeController extends Controller
         // Nối id tới secret ứng mỗi user khác nhau
         $secret = new Secret;
         $user = Auth::user();
-        $secret->user_id = $user->id;
-        $secret->note_id = $note->id;
+        $secret->owner_id = $user->id;
+        $secret->asset_id = $note->id;
         // TODO: encrypt OpenGPG
         $secret->data = $req->cipher;
         $secret->save();
@@ -309,7 +328,7 @@ class HomeController extends Controller
         $note->save();
 
         if ($request->cipher) {
-            $secret = Secret::where('note_id', $note_id)->first();
+            $secret = Secret::where('asset_id', $note_id)->first();
             $secret->data = $request->cipher;
             $secret->save();
         }
@@ -328,7 +347,7 @@ class HomeController extends Controller
         $note = Note::find($note_id);
         $note->delete();
 
-        $secret = Secret::where('note_id', $note_id)->first();
+        $secret = Secret::where('asset_id', $note_id)->first();
         $secret->delete();
 
         $notes = Auth::user()->note()->get();
@@ -416,17 +435,17 @@ class HomeController extends Controller
         return Storage::disk('userstorage')->download(Auth::user()->id.'/'.$filename);
     }
 
-    public function sendMail()
-    {
-        $data = array('name'=>"Ngan", "body" => "Test mail");
+    // public function sendMail()
+    // {
+    //     $data = array('name'=>"Ngan", "body" => "Test mail");
     
-        Mail::send('emails.sendmail', $data, function($message) {
-            $message->to('dangthingan1996@gmail.com', 'Ngan Dang')
-                    ->subject('Web Testing Mail');
-            $message->from('ngandt52@gmail.com','SecPass');
-        });
-        return "Your email has been sent successfully";
-    }
+    //     Mail::send('emails.sendmail', $data, function($message) {
+    //         $message->to('dangthingan1996@gmail.com', 'Ngan Dang')
+    //                 ->subject('Web Testing Mail');
+    //         $message->from('ngandt52@gmail.com','SecPass');
+    //     });
+    //     return "Your email has been sent successfully";
+    // }
     
     public function sharewithme()
     {
@@ -435,9 +454,93 @@ class HomeController extends Controller
         $accounts = Account::whereIn('id', $asset_ids)->get();
         $notes = Note::whereIn('id', $asset_ids)->get();
 
-        return view('page.sharewithme', compact('assets','accounts','notes'));
+        return view('page.sharewithme', compact('accounts','notes'));
     }
 
+    public function deleteAsset(Request $request)
+    {
+        $asset_id = $request->id;
+        Share::where('asset_id', $asset_id)->delete();
+        Secret::where('asset_id', $asset_id)->delete();
+
+        Account::where('id', $asset_id)->delete();
+        Note::where('id', $asset_id)->delete();
+
+        $assets = Auth::user()->share()->get();
+        $asset_ids = $assets->pluck('asset_id');
+        $accounts = Account::whereIn('id', $asset_ids)->get();
+        $notes = Note::whereIn('id', $asset_ids)->get();
+
+        return response()->json([
+            'success' => true,
+            // TODO: lang this message
+            'message' => 'Xoá thành công',
+            'view' => view('content.content-sharewithme', compact('accounts','notes'))->render()
+        ]);
+    }
+
+    public function moveAsset(Request $request)
+    {
+        $asset_id = $request->id;
+        Share::where('asset_id', $asset_id)->delete();
+
+        $assets = Auth::user()->share()->get();
+        $asset_ids = $assets->pluck('asset_id');
+        $accounts = Account::whereIn('id', $asset_ids)->get();
+        $notes = Note::whereIn('id', $asset_ids)->get();
+
+        return response()->json([
+            'success' => true,
+            // TODO: lang this message
+            'message' => 'Đã chuyển thành công',
+            'view' => view('content.content-sharewithme', compact('accounts','notes'))->render()
+        ]);
+    }
+
+    public function moveAccounts(Request $request)
+    {
+        $assets = Auth::user()->share()->get();
+        $asset_ids = $assets->pluck('asset_id');
+
+        $accounts = Account::whereIn('id', $asset_ids)->get();
+        $account_ids = $accounts->pluck('id');
+        Share::whereIn('asset_id', $account_ids)->delete();
+
+        $assets = Auth::user()->share()->get();
+        $asset_ids = $assets->pluck('asset_id');
+        $accounts = Account::whereIn('id', $asset_ids)->get();
+        $notes = Note::whereIn('id', $asset_ids)->get();
+
+        return response()->json([
+            'success' => true,
+            // TODO: lang this message
+            'message' => 'Đã chuyển thành công',
+            'view' => view('content.content-sharewithme', compact('accounts','notes'))->render()
+        ]);        
+    }
+    
+    public function moveNotes(Request $request)
+    {
+        $assets = Auth::user()->share()->get();
+        $asset_ids = $assets->pluck('asset_id');
+
+        $notes = Note::whereIn('id', $asset_ids)->get();
+        $note_ids = $notes->pluck('id');
+        Share::whereIn('asset_id', $note_ids)->delete();
+
+        $assets = Auth::user()->share()->get();
+        $asset_ids = $assets->pluck('asset_id');
+        $accounts = Account::whereIn('id', $asset_ids)->get();
+        $notes = Note::whereIn('id', $asset_ids)->get();
+
+
+        return response()->json([
+            'success' => true,
+            // TODO: lang this message
+            'message' => 'Đã chuyển thành công',
+            'view' => view('content.content-sharewithme', compact('accounts','notes'))->render()
+        ]);
+    }
 
     public function groups()
     {
@@ -452,9 +555,10 @@ class HomeController extends Controller
                 ->join('users', 'groups_users.user_id', '=', 'users.id')
                 ->select('groups_users.*','users.email','users.name')
                 ->get();
-        $admin = Auth::user()->GroupUser()->where('group_id',$group->id)->first()->is_admin;
-        
-        return view('page.groupdetail', compact('group','groupUsers','admin'));
+        $admin = Auth::user()->GroupUser()->where('group_id', $group->id)->first()->is_admin;
+        $accounts = $group->account()->get();
+        $notes = $group->note()->get();
+        return view('page.groupdetail', compact('group','groupUsers','admin', 'accounts', 'notes'));
     }
     public function checkUser(Request $request)
     {
@@ -494,24 +598,17 @@ class HomeController extends Controller
        
         $data = json_decode(stripslashes($_POST['li_variable']));
 
-        foreach($data as $d){
-            $user = User::where('email',$d)->first();
-            $group_user = new GroupUser;
-            $group_user->group_id = $group->id;
-            $group_user->user_id = $user->id;
-            $group_user->save();
+        foreach($data as $email){
+            $user = User::where('email',$email)->first();
+            if( $user != Auth::user() ) 
+            {
+                $group_user = new GroupUser;
+                $group_user->group_id = $group->id;
+                $group_user->user_id = $user->id;
+                $group_user->save();          
+            }
         }
         
-        // $users = User::where('email',$request->email)->get();
-        // foreach ($list as $email)
-        // {
-        //     $user = User::where('email',$email)->first();
-
-        //     $group_user = new GroupUser;
-        //     $group_user->group_id = $group->id;
-        //     $group_user->user_id = $user->id;
-        //     $group_user->save();
-        // }
         $groups = Auth::user()->group()->get();
         return response()->json([
             'success' => true,
@@ -580,12 +677,23 @@ class HomeController extends Controller
     }
     public function deleteUser(Request $request)
     {
-        $user_id = $request->idDelete;
-        $group_id = $request->idGroup;
+        $user_id = $request->user_id;
+        $group_id = $request->group_id;
         $group_user = GroupUser::where('group_id',$group_id)
                                 ->where('user_id', $user_id)->first();
+        $group = GroupUser::where('group_id',$group_id)->get();
+
+        if( ($group_user->is_admin == true) && ( count($group->where('is_admin',true)) == 1) ) {
+            return response()->json([
+                'success' => false,
+                // TODO: lang this message
+                'message' => 'Vui lòng chọn quản trị viên thay thế'
+            ],'500');
+        }
+
         $group_user->delete();
 
+        // Refresh content
         $group = Group::find($group_id);
         $groupUsers = DB::table('groups_users')
                 ->where('group_id', $group_id)        
@@ -597,7 +705,7 @@ class HomeController extends Controller
         return response()->json([
             'success' => true,
             // TODO: lang this message
-            'message' => 'Xóa người dùng thành công',
+            'message' => 'Xóa người dùng khỏi nhóm thành công',
             'view' => view('content.content-group-user', compact('group','groupUsers','admin'))->render()
         ]);
 
@@ -625,8 +733,6 @@ class HomeController extends Controller
                 $user->is_admin = 0;
             }
         }
-        // $group_user->group_id = $group->id;
-        // $group_user->user_id = $user->id;
         $user->save();
 
         $group = Group::find($group_id);
@@ -651,19 +757,45 @@ class HomeController extends Controller
         return view('page.profile', compact('user'));
     }
 
+    public function saveProfile(Request $request)
+    { 
+        $user = Auth::user();
+        $user->name = $request->name;
+        $user->save();
+
+        $profile = $user->profile();
+        $profile->avatar = $request->avatar;
+        $profile->firstname = $request->firstname;
+        $profile->lastname = $request->lastname;
+        $profile->gender = $request->gender;
+        $profile->date_of_birth = $request->date_of_birth;
+        $profile->phone = $request->phone;
+        $profile->address = $request->address;
+        $profile->timezone = $request->timezone;
+        $profile->language = $request->language;
+        $profile->save();
+
+        return response()->json([
+            'success' => true,
+            // TODO: lang this message
+            'message' => 'Lưu thay đổi thành công',
+            'view' => view('content.profile', compact('user'))->render()
+        ]);
+    }
+
     public function addPrivKey(Request $request)
     { 
         try {            
             $user = Auth::user();
             // Only allow 2 keys in early development
             $pgp = PGPkey::where([
-                ['user_id', $user->id],
+                ['owner_id', $user->id],
                 ['type', '5'],
             ]);
             $pgp->delete();
 
             $pgp_key = new PGPkey;
-            $pgp_key->user_id = $user->id;
+            $pgp_key->owner_id = $user->id;
             $pgp_key->armored_key = $request->armored_key;
             $pgp_key->uid = $request->uid;
             $pgp_key->key_id = $request->key_id;
@@ -703,7 +835,7 @@ class HomeController extends Controller
         // Delete private key(s) as user requested
         $user = Auth::user();
         $pgp = PGPkey::where([
-            ['user_id', $user->id],
+            ['owner_id', $user->id],
             ['type', '5'],
         ]);
         $pgp->delete();
@@ -722,11 +854,13 @@ class HomeController extends Controller
         
         return view('content.quicksearch', compact('accounts','notes','groups'));
     } 
+
     public function getUser()
     {
         $users = User::where('role_id','!=','5bdf5220-d75c-11e8-843b-a7f6cbee423d')->get();
         return view('admin.user-manage', compact('users'));
     }
+
     public function editUser(Request $request)
     {
         $id = $request->id;
@@ -746,11 +880,13 @@ class HomeController extends Controller
             'view' => view('admin.content-user-manage', compact('users'))->render()
         ]);
     }
+
     public function getGroup()
     {
         $groups = Group::all();
         return view('admin.group-manage', compact('groups'));
     }
+    
     public function pgp()
     {
         return view('page.pgp');
